+++ src/Accounts/Authentication/Sanitizer/Providers/SanitizerCollectionProvider.cs
@@ -39,13 +39,16 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Sanitizer.Providers
                         var collItemType = collItem.GetType();
                         if (collItemType == typeof(string))
                         {
-                            if (Service.TrySanitizeData(collItem as string, out string sanitizedData))
+                            if (Service.TrySanitizeData(collItem as string, out var detections, out _))
                             {
                                 telemetry.SecretsDetected = true;
                                 var propertyPath = ResolvePropertyPath(property);
                                 if (!string.IsNullOrEmpty(propertyPath))
                                 {
-                                    telemetry.DetectedProperties.Add(ResolvePropertyPath(property));
+                                    foreach (var detection in detections)
+                                    {
+                                        telemetry.DetectedProperties.AddPropertyInfo(propertyPath, detection.Moniker);
+                                    }
                                 }
                             }
                         }