+++ src/Network/Network/ChangeLog.md
@@ -20,6 +20,20 @@
 
 ## Upcoming Release
 
+## Version 7.10.0
+* Onboarded Azure Virtual Network Manager Cmdlets for UDR and NSG Management
+    - `New/Get/Remove/Set-AzNetworkManagerRoutingConfiguration`
+    - `New/Get/Remove/Set-AzNetworkManagerRoutingRuleCollection`
+    - `New/Get/Remove/Set-AzNetworkManagerRoutingRule`
+    - `New-AzNetworkManagerRoutingGroupItem`
+    - `New-AzNetworkManagerRoutingRuleDestination`
+    - `New-AzNetworkManagerRoutingRuleNextHop`
+    - `New/Get/Remove/Set-AzNetworkManagerSecurityUserConfiguration`
+    - `New/Get/Remove/Set-AzNetworkManagerSecurityUserRuleCollection`
+    - `New/Get/Remove/Set-AzNetworkManagerSecurityUserRule`
+    - `New-AzNetworkManagerSecurityUserGroupItem`
+* Added support for `MemberType` property in `New-AzNetworkManagerSecurityUserGroupItem` command
+
 ## Version 7.9.0
 * Onboarded `Microsoft.VideoIndexer/accounts` to private link cmdlets
 * Added support to create, get and delete Bastion shareable links
@@ -35,8 +49,9 @@
 * Added support of `DestinationIPAddress` property in `New-AzPrivateLinkService` command
     - `LoadBalancerFrontendIpConfiguration` is not a mandatory parameter anymore.
     - The user can provide either `LoadBalancerFrontendIpConfiguration` or `DestinationIPAddress`.
-* Added optional parameter 'ResiliencyModel' to Express Route Virtual Network Gateway
-    
+* Added support for `PrivateEndpointVNetPolicies` property in `New-AzVirtualNetwork`
+* Added support for `ResiliencyModel` property in `New-AzVirtualNetworkGateway`
+
 ## Version 7.8.1
 * Fixed secrets exposure in example documentation.
 * Onboarded `Microsoft.App/managedEnvironments` to private link cmdlets